+++ app/hooks.py
@@ -1,19 +1,39 @@
-from app.managers.entity_manager import EntityManager
-from app.managers.cache_manager import CacheManager
+"""
+This module defines a system for executing various hooks based on
+specific actions within the application. It includes the Hook class,
+which orchestrates post-event operations by interacting with
+EntityManager and CacheManager. The H enumeration specifies different
+hook types, and the Hook class manages their execution, handling actions
+related to user management, collections, documents, comments, downloads,
+and favorites. This setup enables asynchronous processing and integrates
+seamlessly with session and caching systems to ensure efficient state
+management and responsiveness to application events.
+"""
+
 import enum
-from app.context import get_context
+from typing import Any
 from redis import Redis
 from sqlalchemy.ext.asyncio import AsyncSession
-from app.models.user_models import User
-from typing import Any
 from fastapi import Request
+from app.managers.entity_manager import EntityManager
+from app.managers.cache_manager import CacheManager
+from app.models.user_models import User
+from app.context import get_context
 
 ctx = get_context()
 
 
 class H(enum.Enum):
+    """
+    Defines an enumeration of hook types used to manage various
+    post-event operations within the application. Each item of this
+    enumeration represents a specific action or event, such as user
+    registration, document upload, or comment update. These hook types
+    are used to trigger corresponding functions that handle tasks
+    related to user management, collections, documents, comments,
+    downloads, and favorites.
+    """
     AFTER_STARTUP = "after_startup"
-
     AFTER_USER_REGISTER = "after_user_register"
     AFTER_USER_LOGIN = "after_user_login"
     AFTER_TOKEN_RETRIEVE = "after_token_retrieve"
@@ -25,26 +45,21 @@ class H(enum.Enum):
     AFTER_USERPIC_UPLOAD = "after_userpic_upload"
     AFTER_USERPIC_DELETE = "after_userpic_delete"
     AFTER_USERS_LIST = "after_users_list"
-
     AFTER_COLLECTION_INSERT = "after_collection_insert"
     AFTER_COLLECTION_SELECT = "after_collection_select"
     AFTER_COLLECTION_UPDATE = "after_collection_update"
     AFTER_COLLECTION_DELETE = "after_collection_delete"
     AFTER_COLLECTIONS_LIST = "after_collections_list"
-
     AFTER_DOCUMENT_UPLOAD = "after_document_upload"
     AFTER_DOCUMENT_DOWNLOAD = "after_document_download"
     AFTER_DOCUMENT_SELECT = "after_document_select"
-
     AFTER_COMMENT_INSERT = "after_comment_insert"
     AFTER_COMMENT_SELECT = "after_comment_select"
     AFTER_COMMENT_UPDATE = "after_comment_update"
     AFTER_COMMENT_DELETE = "after_comment_delete"
     AFTER_COMMENTS_LIST = "after_comments_list"
-
     AFTER_DOWNLOAD_SELECT = "after_download_select"
     AFTER_DOWNLOADS_LIST = "after_downloads_list"
-
     AFTER_FAVORITE_INSERT = "after_favorite_insert"
     AFTER_FAVORITE_SELECT = "after_favorite_select"
     AFTER_FAVORITE_DELETE = "after_favorite_delete"
@@ -52,15 +67,39 @@ class H(enum.Enum):
 
 
 class Hook:
+    """
+    Manages and executes various hooks for handling post-event
+    operations within the application. This class initializes with
+    necessary components such as an entity manager, cache manager,
+    request, and current user. The execute method runs the appropriate
+    hook functions based on the specified hook action, processing data
+    as required and returning the result.
+    """
 
     def __init__(self, session: AsyncSession, cache: Redis,
                  request: Request = None, current_user: User = None):
+        """
+        Initializes the Hook class with an entity manager,
+        cache manager, request, and current user. The entity manager
+        is created from the provided session, and the cache manager is
+        created from the provided Redis instance. The request and
+        current user are optional and can be used to provide context
+        for the hook execution.
+        """
         self.entity_manager = EntityManager(session)
         self.cache_manager = CacheManager(cache)
         self.request = request
         self.current_user = current_user
 
     async def execute(self, hook_action: H, data: Any = None) -> Any:
+        """
+        Executes the specified hook action by calling the associated
+        functions with the provided entity manager, cache manager,
+        request, current user, and data. The hook functions are
+        retrieved from the context based on the hook action value and
+        are invoked sequentially. Returns the processed data after
+        executing all hook functions.
+        """
         if hook_action.value in ctx.hooks:
             hook_functions = ctx.hooks[hook_action.value]
             for func in hook_functions: