+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+Version 0.0.29 (2024-08-25)
+-------------
+
+- Added a new template extension for managing application hooks.
+- Added lifecycle hooks for user actions, collection operations, and
+  document management.
+
 Version 0.0.28 (2024-08-25)
 -------------